+++ packages/typography/src/lib/Typography.tsx
@@ -1,6 +1,24 @@
 import { forwardRef } from 'react';
 import { default as MuiTypography, TypographyProps as MuiTypographyProps } from '@mui/material/Typography';
 
+declare module '@mui/material/styles' {
+  interface TypographyVariants {
+    agreement: React.CSSProperties;
+  }
+
+  // allow configuration using `createTheme`
+  interface TypographyVariantsOptions {
+    agreement?: React.CSSProperties;
+  }
+}
+
+// Update the Typography's variant prop options
+declare module '@mui/material/Typography' {
+  interface TypographyPropsVariantOverrides {
+    agreement: true;
+  }
+}
+
 export interface TypographyProps extends MuiTypographyProps {
   children?: React.ReactNode;
 }