+++ packages/datepicker/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.4.11](https://github.com/Availity/element/compare/@availity/mui-datepicker@0.4.10...@availity/mui-datepicker@0.4.11) (2024-06-14)
+
+### Dependency Updates
+
+* `mui-textfield` updated to version `0.4.10`
+* `theme-provider` updated to version `0.4.10`
+## [0.4.10](https://github.com/Availity/element/compare/@availity/mui-datepicker@0.4.9...@availity/mui-datepicker@0.4.10) (2024-06-12)
+
+### Dependency Updates
+
+* `theme-provider` updated to version `0.4.9`
 ## [0.4.9](https://github.com/Availity/element/compare/@availity/mui-datepicker@0.4.8...@availity/mui-datepicker@0.4.9) (2024-05-17)
 
 ### Dependency Updates