+++ packages/element/CHANGELOG.md
@@ -2,6 +2,51 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.40.3](https://github.com/Availity/element/compare/@availity/element@0.40.2...@availity/element@0.40.3) (2024-06-17)
+
+### Dependency Updates
+
+* `mui-spaces` updated to version `0.40.2`
+## [0.40.2](https://github.com/Availity/element/compare/@availity/element@0.40.1...@availity/element@0.40.2) (2024-06-14)
+
+### Dependency Updates
+
+* `mui-accordion` updated to version `0.40.1`
+* `mui-authorize` updated to version `0.40.1`
+* `mui-autocomplete` updated to version `0.40.1`
+* `mui-block-ui` updated to version `0.40.1`
+* `mui-checkbox` updated to version `0.40.1`
+* `mui-datepicker` updated to version `0.40.1`
+* `mui-disclaimer` updated to version `0.40.1`
+* `mui-feedback` updated to version `0.40.1`
+* `mui-form-utils` updated to version `0.40.1`
+* `mui-modal` updated to version `0.40.1`
+* `mui-page-header` updated to version `0.40.1`
+* `mui-popover` updated to version `0.40.1`
+* `mui-progress` updated to version `0.40.1`
+* `mui-spaces` updated to version `0.40.1`
+* `mui-table` updated to version `0.40.1`
+* `mui-textfield` updated to version `0.40.1`
+* `mui-typography` updated to version `0.40.1`
+* `theme-provider` updated to version `0.40.1`
+## [0.40.1](https://github.com/Availity/element/compare/@availity/element@0.40.0...@availity/element@0.40.1) (2024-06-12)
+
+### Dependency Updates
+
+* `mui-checkbox` updated to version `0.40.0`
+* `mui-page-header` updated to version `0.40.0`
+## [0.40.0](https://github.com/Availity/element/compare/@availity/element@0.39.3...@availity/element@0.40.0) (2024-06-12)
+
+### Dependency Updates
+
+* `mui-datepicker` updated to version `0.39.3`
+* `mui-disclaimer` updated to version `0.39.3`
+* `theme-provider` updated to version `0.39.3`
+
+### Features
+
+* **mui-disclaimer:** add disclaimer component ([841e989](https://github.com/Availity/element/commit/841e98900adc535171fc82865a931d802ddab96c))
+
 ## [0.39.3](https://github.com/Availity/element/compare/@availity/element@0.39.2...@availity/element@0.39.3) (2024-06-10)
 
 ### Dependency Updates