+++ packages/table/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.11](https://github.com/Availity/element/compare/@availity/mui-table@0.1.10...@availity/mui-table@0.1.11) (2024-06-14)
+
+### Dependency Updates
+
+* `mui-form-utils` updated to version `0.1.10`
+* `mui-textfield` updated to version `0.1.10`
+* `mui-typography` updated to version `0.1.10`
 ## [0.1.10](https://github.com/Availity/element/compare/@availity/mui-table@0.1.9...@availity/mui-table@0.1.10) (2024-06-10)
 
 ### Dependency Updates